Remove compute method from LinkCell, everywhere in tests
