+++ cpp/locality/LinkCell.cc
@@ -164,93 +164,6 @@ void LinkCell::computeCellList(const box::Box& box, const vec3<float>* points, u
     }
 }
 
-void LinkCell::compute(const box::Box& box, const vec3<float>* points, unsigned int n_points,
-                       const vec3<float>* query_points, unsigned int n_query_points, bool exclude_ii)
-{
-    // Store points ("j" particles in (i, j) bonds) in the cell list
-    // for quick access later (not points)
-    computeCellList(box, points, n_points);
-
-    typedef std::vector<NeighborBond> BondVector;
-    typedef std::vector<BondVector> BondVectorVector;
-    typedef tbb::enumerable_thread_specific<BondVectorVector> ThreadBondVector;
-    ThreadBondVector bond_vectors;
-
-    // Find (i, j) neighbor pairs
-    parallel_for(blocked_range<size_t>(0, n_query_points), [=, &bond_vectors](const blocked_range<size_t>& r) {
-        ThreadBondVector::reference bond_vector_vectors(bond_vectors.local());
-        bond_vector_vectors.emplace_back();
-        BondVector& bond_vector(bond_vector_vectors.back());
-
-        for (size_t i(r.begin()); i != r.end(); ++i)
-        {
-            // get the cell the point is in
-            const vec3<float> point(query_points[i]);
-            const unsigned int point_cell(getCell(point));
-
-            // loop over all neighboring cells
-            const std::vector<unsigned int>& neigh_cells = getCellNeighbors(point_cell);
-            for (unsigned int neigh_idx = 0; neigh_idx < neigh_cells.size(); neigh_idx++)
-            {
-                const unsigned int neigh_cell = neigh_cells[neigh_idx];
-
-                // iterate over the particles in that cell
-                locality::LinkCell::iteratorcell it = itercell(neigh_cell);
-                for (unsigned int j = it.next(); !it.atEnd(); j = it.next())
-                {
-                    if (exclude_ii && i == j)
-                        continue;
-
-                    const vec3<float> rij(m_box.wrap(points[j] - point));
-                    const float rsq(dot(rij, rij));
-
-                    if (rsq < m_cell_width * m_cell_width)
-                    {
-                        bond_vector.emplace_back(i, j, sqrt(rsq));
-                    }
-                }
-            }
-        }
-    });
-
-    // Sort neighbors by particle i index
-    tbb::flattened2d<ThreadBondVector> flat_bond_vector_groups = tbb::flatten2d(bond_vectors);
-    BondVectorVector bond_vector_groups(flat_bond_vector_groups.begin(), flat_bond_vector_groups.end());
-    tbb::parallel_sort(bond_vector_groups.begin(), bond_vector_groups.end(), compareFirstNeighborPairs);
-
-    unsigned int num_bonds(0);
-    for (BondVectorVector::const_iterator iter(bond_vector_groups.begin()); iter != bond_vector_groups.end();
-         ++iter)
-        num_bonds += iter->size();
-
-    m_neighbor_list.resize(num_bonds);
-    m_neighbor_list.setNumBonds(num_bonds, n_query_points, n_points);
-
-    size_t* neighbor_array(m_neighbor_list.getNeighbors());
-    float* neighbor_weights(m_neighbor_list.getWeights());
-    float* neighbor_distances(m_neighbor_list.getDistances());
-
-    // build nlist structure
-    parallel_for(blocked_range<size_t>(0, bond_vector_groups.size()),
-                 [=, &bond_vector_groups](const blocked_range<size_t>& r) {
-                     size_t bond(0);
-                     for (size_t group(0); group < r.begin(); ++group)
-                         bond += bond_vector_groups[group].size();
-
-                     for (size_t group(r.begin()); group < r.end(); ++group)
-                     {
-                         const BondVector& vec(bond_vector_groups[group]);
-                         for (BondVector::const_iterator iter(vec.begin()); iter != vec.end(); ++iter, ++bond)
-                         {
-                            neighbor_array[2 * bond] = iter->id;
-                            neighbor_array[2 * bond + 1] = iter->ref_id;
-                            neighbor_weights[bond] = iter->weight;
-                            neighbor_distances[bond] = iter->distance;
-                         }
-                     }
-                 });
-}
-
 const std::vector<unsigned int>& LinkCell::computeCellNeighbors(unsigned int cur_cell)
 {
     std::vector<unsigned int> neighbor_cells;
