+++ tests/test_interface.py
@@ -47,6 +47,7 @@ class TestInterface(unittest.TestCase):
     def test_filter_r(self):
         """Test that nlists are filtered to the correct rmax."""
         np.random.seed(0)
+        r_max = 3.0
         (box, positions) = util.make_fcc(4, 4, 4, noise=1e-2)
 
         index = np.random.randint(0, len(positions))
@@ -55,11 +56,12 @@ class TestInterface(unittest.TestCase):
         others = np.concatenate([positions[:index], positions[index + 1:]])
 
         # Creates a neighborlist with rmax larger than the interface rmax
-        lc = freud.locality.LinkCell(box, 3.0).compute(box, others, point)
+        lc = freud.locality.LinkCell(box, r_max, others)
+        nlist = lc.query(point, dict(r_max=r_max)).toNList()
 
         inter = freud.interface.InterfaceMeasure(1.5)
 
-        test_twelve = inter.compute(box, others, point, lc.nlist)
+        test_twelve = inter.compute(box, others, point, nlist)
         self.assertEqual(test_twelve.point_count, 12)
         self.assertEqual(len(test_twelve.point_ids), 12)
